Create a popup widget for notifications
This should be the same or super close to the existing notification popup widget in use.

This also adds tracking in the DBus server for active popups.

The idea for this is that there is a central place (here) to show notification popups. That way, Raven can focus on its own view for stored notifications without having to deal with the popup logic.

Hopefully, Raven will somehow be able to use the new notification class without having to duplicate a ton of logic.