Rip out all of the old Raven notification handling
This marks the first time that the new notification system can actually work. Many initial issues with the popups and popup system have already been fixed as a result of this.

There seems to be some weirdness with certain notifications, primarily from Discord, where the notifications will appear on top of each other.