Fix exclusion of valid applications.
Because a window type returned can be a list of types, we should actually check that the window has in its list a _NET_WM_WINDOW_TYPE_NORMAL or _NET_WM_WINDOW_TYPE_TOOLBAR.

If this type is *in* the types, then we should allow this item to be tracked, otherwise ignore. This fixes cases with some KDE applications like Spectacle and KColorChooser not showing up.

Renamed to is_allowed_window_type to better reflect intent with function.