Fix some applications showing their own notification popups when paused
Before this commit, Firefox would show its own notification popups whenever Budgie notifications or popups were paused or disabled. This seems to fix that issue: when notifications are paused, no popups of any kind are shown.

I have no idea why this matters to Firefox, nor how it even detects it, but so far it seems to be working.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>