Fix inconsistency when opening Raven
Per the answer here (https://stackoverflow.com/questions/22178524/gtk-named-stack-childs), if a stack child is not visible, asking the stack to set the visible child will silently fail, apparently even if the child being shown *is* visible.

The NotificationView widget already calls show_all() in its constructor, so the call only had to be added to the applet window. This fixes the case where clicking the notification bell fails to open the correct page, and possibly also fixes where clicking the Raven trigger for the first time sometimes fails to open Raven (it is technically open, but the elements arent visible).

Ref #31

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>