Remove delay when opening apps
This was introduced in Abomination in an attempt to fix Chrome profile manager closing sequence but endend up breaking some other apps (such as Steam) closing sequence.
Instead, try to reparent first app of group when an app associated with a pinned icon is closed so that the pinned button is reused.