Drop Calendar menu item from ClockApplet.
Unfortunately AppInfo get_default_for_type had the tendency to provide applications that did not really handle text/calendar, such as gedit, resulting in user confusion. Since the underlying struct of DesktopAppInfo and AppInto do not expose MimeType, we cannot perform validation of the mimetypes to dynamically hide the button. So, just remove it.

This may make a return in B11 when we are no longer relying on glib2 for this behaviour.