Move the application index and tracking to its own module
This allows the system to be reused in multiple places, e.g. the Budgie Menu and the F2 Launcher.

To ensure we only have a single copy of the index, the category list has been made static, ensuring that we aren't wasting a bunch of memory and resources on multiple copies of the index.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>