Set a miminum expiry timeout for notification popups
Some applications set a timeout of 0, which the spec says means that it should never expire. Our implementation doesn't really fit that, since persistence is done by Raven, not the notification server. So, we should have a minimum show time for the popups to ensure that one is actually shown when it should be.

Fixes #129

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>