Change label properties for notification bodies (#179)
This makes notificiation popup bodies more consistent, and prevents them from taking up unnecessarily large portions of the screen.

The max chars was reduced slightly to ensure that the close button doesn't get shoved off the edge of the notification popup. Width chars was added so that the body text always has the same minimum and maximum size per line, so the entire popup body is filled by the label. Lines is set to 2 so only two lines of text is displayed. Note that this does not mean two lines in the popup; it's two lines of the actual notification text. This allows long lines to wrap (and wrapped lines don't count towards lines, because reasons) without popups taking up large amounts of screen real-estate. Xalign is set to 0 to ensure that the text always starts at the left (because halign and justify isn't enough).

Ref #171

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>