Run Dialog refactor and improvements
This implements using the new application indexer used by the new Budgie Menu to find and sort applications. When searching, applications are filtered and sorted just like the Budgie menu taking fuzzy finding and relevancy into account.

To make sure that the Run Dialog closes as expected in all cases, the context signals for launching an application were hooked up and re-emitted. This allows the Run Dialog to use the same function for launching applications as the Budgie Menu, which enables setting GPU profiles if necessary for multi-GPU systems.

There a a couple of visual changes as well:

1. The width is now computed from the workarea of the screen that the Run Dialog is opened on. This should make it look good on different sized monitors.

2. The label for the name and description of applications in the launcher buttons has been split into two labels. This allows more fine-grained styling than was previously possible.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>