Fix large fonts pushing the close button out of the popup
This sets the notification body to use the size of the widget for wrapping instead of a character length. Notification titles also were given a margin-right to prevent overlapping the close button.

Speaking of the close button, it is now overlayed over the body with a GtkOverlay widget. This allows all of the close handlers to be done by the base popup widget instead of each part of the widget re-broadcasting a signal that the close button was clicked.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>