Code cleanup and tweaks to Screenshot implementation.
This commit implements some code cleanup to the implementation of Budgie Screenshot, reducing some conditional blocks in favor of ternaries as an example, early returns, etc. Removed a bunch of random "don't translate" comments (they wouldn't be translated because it wasn't using the right gettext syntax anyways).

Alongside this, Budgie Screenshot has received various changes:

1. It is a normal window now instead of a utility so it can be better tracked and favorited for the IconTasklist, like it would be if it was more of an external application akin to GNOME Screenshot.
2. Previously you could not cancel area selection. You can now do this with right-click. Ideally in the future key-based cancellation with Escape will work too.
3. Various window class changes and renamed from budgie-screenshot to the standardized reverse DNS name of org.buddiesofbudgie.Screenshot.

IconTasklist has received some workarounds for Budgie Screenshot. As the implementation is invoked by budgie-daemon, for whatever reason GTK still sets its X window information to that (even when corrected with startup_id and wmclass). So we do renaming in IconTasklist. This does not fix it for the legacy Tasklist. Patches welcome.