fix: Inability to capture fullscreen applications like games.
Fixes #298. This also cleans up some duplication of logic between screen-type screenshot and area, where screen is just an area of the dimensions of the screen.

Similar to GNOME Shell, we will dynamically disable undirect, before potentially re-enabling it (if it was not explicitly disabled in our settings before the call to screenshot).

Tested in the following games, all running in exclusive fullscreen mode, across several instances of both "screen" and "window" capture to validate that these aren't just one-offs that worked:

- Halo: Reach in MCC
- OpenMW
- Sid Meier's Civilization VI

All the above games were running in exclusive fullscreen modes.