Modify array_contains to try to prevent better scores by mistake
Checking if the search term is just anywhere in the text string tends to leads to applications having better scores than intended, for instance making the Power panel of the Control Center appear above Terminal when searching for "ter".

Also, we were seeing if the text string matched the search term when logically it should be the other way around.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>