simple-tasklist: Let Pango figure out ellipsizing
This has the strange side-effect of causing the applet section that the applet is in to no longer start at the... well, start of the panel, assuming the applet is in the start section of the panel. Instead, the section becomes centered. I have never seen this happen before and have no idea why it does this. Yet.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>