polkit-dialog: Add a revealer to show details about the escallation
This will show the action description and action ID in the Polkit authentication dialog.

To fascilitate this, the VAPI files for Polkit had to be updated/fixed since the method signatures for the needed functions were not correct. Debian is still on version 105 (:screaming_face:) while I have 122, but looking at the Polkit repository at the 105 tag still has the same functions with the same signatures, so we should be good there.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>