simple-tasklist: Migrate from Abomination to libxfce4windowing
The end goal (likely in another branch and this will be rebased) is for Abomination to wrap libxfce4windowing so it can be used for both tasklist applets. This work, though, gets me familiar with the library and its capabilities. My learnings from this can then be used in Abomination.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>