notification_view: Simplify handling of notification groups in Raven
This does a couple things. First, when constructing a notification group, just get the image directly from our Notification object. We did all the work to get the correct image, let's actually use it.

Secondly, remove our HashTable mapping application name to notification group. We don't need it; anything it was used for we can get from the ListBox. This avoids having to keep track of groups in multiple data structures (the table and the list box). This makes the code easier to maintain, and slightly more memory effecient.

Lastly, switch to `ngettext` to build the text for the notification view header instead of an if/elseif chain.

Fixes #450

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>