RFC: Initial pass at a CODEOWNERS
This pull request is an RFC to adds codeowner defaults for various parts of the codebase, based on historical or regular contributions. It is intended to be used to enforce or mandate blocking individuals for merge. Listings can / should be modified based on interest / desire for code reviews, not necessarily considered exhaustive.

All listings have an addition of the Best Buds team, so anyone on the team should be notified automatically. Global rule notifies the Best Buds team as well.