notifications: Try to guess app id if getting from hints fails
The notification spec states that the `desktop-entry` hint should be the prefix of the Desktop Application Info file. However, because of the push to have fully-qualified app ID's, some applications now have a mismatch between the ID and the desktop file, meaning the hint cannot be used for that application.

In cases where the hint fails, try to guess the ID using the application's name given to us from DBus.

It seems to fix #477.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>