icons: Use the correct name for the reboot icon (#584)
According to the icon naming spec, the correct name to use for a reboot icon is `system-reboot` (and thus, `system-reboot-symbolic` for symbolic). Before this, we were using `system-restart-symbolic`, and had to provide an icon for this. This change increases compatibility with other icon themes, including Breeze, while retaining our own icon in case a theme does not provide it.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>