workspace-applet: Fix more issues related to Wayland
The big thing is that the applet no longer crashes when a new window is opened. When a window is opened, it may not yet have an application assigned to it. When that happens, set the item's icon to a generic one, and wait for it to update naturally. In practice, I have yet to see the generic icon; it's always been updated by the time I can see it.