Make the screenshot dialog less aggressive to start when finding libxfce4windowing details. (#710)
We are expecting too much of libxfce4windowing to get the monitor and scale details
on startup - the underlying wayland client needs more time after the desktop has initialised.
This is intended to prevent crashing in the necessary details are not found before first usage.
As a fallback, the autostart session is amended to restart if necessary if at anytime the
dialog fails for any reason.

Taken together this makes the screenshot capability experience more robust