panel: Don't use geometry to detect monitor (#720)
We can't get a window's geometry under Wayland, so instead, try to get
the "primary" monitor from the first Budgie panel we find, and check if
a window's monitor is the same.

It may be janky with workspaces, since we can't yet do
Window<->Workspace association with libxfce4windowing and labwc, but
otherwise, it *seems* to work. Kind of amazing, really.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>