session: Add a start script to start the compositor
This is adapted from Xfce4. The start script sets some environment
variables, copies our configuration files to the correct user directory,
and starts the Wayland compositor, in our case, labwc. They support
running different compositors by passing it as an argument to the
script, which is why they have the whole Opts thing. We may want that in
the future too, so I left it in.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>