Don't read settings after bind
This is actually not needed, bind() already ensures that the right item is
selected in the combobox.

Setting active_id explicitly on load causes that the value is saved back
immediately even if it's not changed by the user, which makes the
session-dependent default value apply also for other sessions.