build: Actually bump budgie-raven-plugin version
With pull request 743, this is now generating budgie-raven-plugin-1.0.pc
and budgie-raven-plugin-2.0.vapi. According to [Vala docs] the VAPI
filename must match the pkgconfig filename which is no longer the case
now.

I am trying to rebuild reverse deps with PR 743 on NixOS and found
e.g. budgie-analogue-clock-applet start failing to build (even after
switching to budgie-2.0) because it failed to find a VAPI file for
budgie-raven-plugin-1.0.

[Vala docs]: https://docs.vala.dev/developer-guides/bindings/writing-a-vapi-manually/02-00-getting-started/02-01-the-vapi-file.html