fix: panel position crash on direct position flip (horizontal or vertical)
There is something gnarly happening either internal;y in GTK show for the toplevels, or with the compositor, that triggers the panel to crash when we attempt to perform a direct position change (e.g. bottom to top).

Fully reconstructing the panel (closing it, removing it from panels hashtable, loading it again) seems to fix it, with the terrible caveat that it breaks conflicting panel position logic.

We'll want to fix duplicate panel position logic in Budgie Desktop Settings later.