settings_panel: Avoid panel conflicts
With 3ac81f6, it is possible to have two panels on the same screen edge.
This causes some very... interesting... (read broken) things to happen,
so we should make sure that scenario isn't possible to create.

This commit does so by checking if a position is already in use by a
panel, and not adding it to the settings model if it is. That means that
we have to store and manage the position model separately so we can
update it when the setting changes.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>